Add files via upload
I edited before Sample Design. We need to choose between the 3 growth variables as our target variable. I now personally think we should pick growth on sales, as it gives the largest sample size given the threshold is set at sample's third quartile, which is the first 25% of the sample data.